Added all necessary files.

+++ UWFNursing/Tests/ATCNetworkTest.m
@@ -31,7 +31,7 @@
     
     [_networkUtils loginUserWithUsername:@"jabbey" andPassword:@"2731" withCompletionHandler:^(NSError *error, NSUInteger userId, NSInteger session, NSString *errorMessage) {
         NSLog(@"Error %@", error);
-        NSLog(@"User id %lu", userId);
+        NSLog(@"User id %d", (int)userId);
         NSLog(@"Error %@", errorMessage);
         
     }];
